// src/plugins/TestPlugin/dataSources/STAClient/api.ts
export interface Datastream {
    /**
     * The Id of the datastream
     * @type {number}
     * @memberof Datastream
     */
    '@iot.id'?: number;
    /**
     * The direct link to the entity
     * @type {string}
     * @memberof Datastream
     */
    '@iot.selfLink'?: string;
    /**
     *
     * @type {Array<Observation>}
     * @memberof Datastream
     */
    'Observations'?: Array<Observation>;
    /**
     * The total number of entities in the entityset
     * @type {number}
     * @memberof Datastream
     */
    'Observations@iot.count'?: number;
    /**
     * A link to a related entity or entity set
     * @type {string}
     * @memberof Datastream
     */
    'Observations@iot.navigationLink'?: string;
    /**
     * The link to the next page of entities
     * @type {string}
     * @memberof Datastream
     */
    'Observations@iot.nextLink'?: string;
    /**
     *
     * @type {ObservedProperty}
     * @memberof Datastream
     */
    'ObservedProperty'?: ObservedProperty;
    /**
     * A link to a related entity or entity set
     * @type {string}
     * @memberof Datastream
     */
    'ObservedProperty@iot.navigationLink'?: string;
    /**
     *
     * @type {Sensor}
     * @memberof Datastream
     */
    'Sensor'?: Sensor;
    /**
     * A link to a related entity or entity set
     * @type {string}
     * @memberof Datastream
     */
    'Sensor@iot.navigationLink'?: string;
    /**
     *
     * @type {Thing}
     * @memberof Datastream
     */
    'Thing'?: Thing;
    /**
     * A link to a related entity or entity set
     * @type {string}
     * @memberof Datastream
     */
    'Thing@iot.navigationLink'?: string;
    /**
     * A description of the datastream
     * @type {string}
     * @memberof Datastream
     */
    'description'?: string;
    /**
     * The label or name of the datastream
     * @type {string}
     * @memberof Datastream
     */
    'name'?: string;
    /**
     * The observation type for the datastream, as defined by the OGC Observations and Measurements standard
     * @type {string}
     * @memberof Datastream
     */
    'observationType'?: string;
    /**
     * The bounding box of the spatial extent of all features of interest associated with observations belonging to the Datastream
     * @type {string}
     * @memberof Datastream
     */
    'observedArea'?: string;
    /**
     * The range of phenomenon times from all observations belonging to the datastream
     * @type {string}
     * @memberof Datastream
     */
    'phenomenonTime'?: string;
    /**
     *
     * @type {DatastreamProperties}
     * @memberof Datastream
     */
    'properties'?: DatastreamProperties;
    /**
     * The range of result times from all observations belonging to the datastream
     * @type {string}
     * @memberof Datastream
     */
    'resultTime'?: string;
    /**
     *
     * @type {DatastreamUnitOfMeasurement}
     * @memberof Datastream
     */
    'unitOfMeasurement'?: DatastreamUnitOfMeasurement;
}
// ...
export interface Location {
    /**
     * The Id of the location
     * @type {number}
     * @memberof Location
     */
    '@iot.id'?: number;
    /**
     * The direct link to the entity
     * @type {string}
     * @memberof Location
     */
    '@iot.selfLink'?: string;
    /**
     *
     * @type {Array<Thing>}
     * @memberof Location
     */
    'Things'?: Array<Thing>;
    /**
     * The total number of entities in the entityset
     * @type {number}
     * @memberof Location
     */
    'Things@iot.count'?: number;
    /**
     * A link to a related entity or entity set
     * @type {string}
     * @memberof Location
     */
    'Things@iot.navigationLink'?: string;
    /**
     * The link to the next page of entities
     * @type {string}
     * @memberof Location
     */
    'Things@iot.nextLink'?: string;
    /**
     * A description of the location
     * @type {string}
     * @memberof Location
     */
    'description'?: string;
    /**
     * The encoding type of the \'location\' property
     * @type {string}
     * @memberof Location
     */
    'encodingType'?: string;
    /**
     * The location, recorded in line with the type defined by the \'encodingType\' property
     * @type {string}
     * @memberof Location
     */
    'location'?: string;
    /**
     * The label or name of the location
     * @type {string}
     * @memberof Location
     */
    'name'?: string;
    /**
     *
     * @type {LocationProperties}
     * @memberof Location
     */
    'properties'?: LocationProperties;
}
// ...
export interface Observation {
    /**
     * The Id of the observation
     * @type {number}
     * @memberof Observation
     */
    '@iot.id'?: number;
    /**
     * The direct link to the entity
     * @type {string}
     * @memberof Observation
     */
    '@iot.selfLink'?: string;
    /**
     *
     * @type {ObservationParameters}
     * @memberof Observation
     */
    'parameters'?: ObservationParameters;
    /**
     * The time when the observation happened
     * @type {string}
     * @memberof Observation
     */
    'phenomenonTime'?: string;
    /**
     * The estimated value of the observed property
     * @type {string}
     * @memberof Observation
     */
    'result'?: string;
    /**
     * A description of the quality of the result
     * @type {string}
     * @memberof Observation
     */
    'resultQuality'?: string;
    /**
     * The time the result was generated
     * @type {string}
     * @memberof Observation
     */
    'resultTime'?: string;
    /**
     * The time period during which the result can be used
     * @type {string}
     * @memberof Observation
     */
    'validTime'?: string;
}
// ...
export interface Observations {
    /**
     * The total number of entities in the entityset
     * @type {number}
     * @memberof Observations
     */
    '@iot.count'?: number;
    /**
     * The link to the next page of entities
     * @type {string}
     * @memberof Observations
     */
    '@iot.nextLink'?: string;
    /**
     *
     * @type {Array<Observation>}
     * @memberof Observations
     */
    'value'?: Array<Observation>;
}
// ...
export interface ObservedProperty {
    /**
     * The Id of the observed property
     * @type {number}
     * @memberof ObservedProperty
     */
    '@iot.id'?: number;
    /**
     * The direct link to the entity
     * @type {string}
     * @memberof ObservedProperty
     */
    '@iot.selfLink'?: string;
    /**
     * A URI for the observed property providing a representation of its definition, ideally through reference to a community based vocabulary
     * @type {string}
     * @memberof ObservedProperty
     */
    'definition'?: string;
    /**
     * A description of the observed property
     * @type {string}
     * @memberof ObservedProperty
     */
    'description'?: string;
    /**
     * The label or name of the observed property
     * @type {string}
     * @memberof ObservedProperty
     */
    'name'?: string;
    /**
     *
     * @type {ObservedPropertyProperties}
     * @memberof ObservedProperty
     */
    'properties'?: ObservedPropertyProperties;
}
// ...
export interface Sensor {
    /**
     * The Id of the sensor type
     * @type {number}
     * @memberof Sensor
     */
    '@iot.id'?: number;
    /**
     * The direct link to the entity
     * @type {string}
     * @memberof Sensor
     */
    '@iot.selfLink'?: string;
    /**
     * A description of the sensor type
     * @type {string}
     * @memberof Sensor
     */
    'description'?: string;
    /**
     * The encoding type of the \'metadata\' property
     * @type {string}
     * @memberof Sensor
     */
    'encodingType'?: string;
    /**
     * A detailed description of the sensor type (style based on \'encodingType\' property)
     * @type {string}
     * @memberof Sensor
     */
    'metadata'?: string;
    /**
     * The label or name of the sensor type
     * @type {string}
     * @memberof Sensor
     */
    'name'?: string;
    /**
     *
     * @type {SensorProperties}
     * @memberof Sensor
     */
    'properties'?: SensorProperties;
}
// ...
export interface Things {
    /**
     * The total number of entities in the entityset
     * @type {number}
     * @memberof Things
     */
    '@iot.count'?: number;
    /**
     * The link to the next page of entities
     * @type {string}
     * @memberof Things
     */
    '@iot.nextLink'?: string;
    /**
     *
     * @type {Array<Thing>}
     * @memberof Things
     */
    'value'?: Array<Thing>;
}
// ...
export class LocationsApi extends BaseAPI implements LocationsApiInterface {
    /**
     * Returns information about the location identified by **entityId**
     * @summary Get information about an individual location
     * @param {string} entityId The id of the requested entity
     * @param {string} [$select] The list of properties that need to be returned
     * @param {string} [$expand] The list of related queries that need to be included in the result
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LocationsApi
     */
    public v11LocationsEntityIdGet(entityId: string, $select?: string, $expand?: string, options?: AxiosRequestConfig) {
        return LocationsApiFp(this.configuration).v11LocationsEntityIdGet(entityId, $select, $expand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Returns information about the thing at the location identified by **entityId**
     * @summary Get information about the thing at the location
     * @param {string} entityId The id of the requested entity
     * @param {number} [$skip] The number of elements to skip from the collection
     * @param {number} [$top] The number of elements to return
     * @param {boolean} [$count] Flag indicating if the total number of items in the collection should be returned.
     * @param {string} [$select] The list of properties that need to be returned
     * @param {string} [$expand] The list of related queries that need to be included in the result
     * @param {string} [$filter] A filter query
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LocationsApi
     */
    public v11LocationsEntityIdThingsGet(entityId: string, $skip?: number, $top?: number, $count?: boolean, $select?: string, $expand?: string, $filter?: string, options?: AxiosRequestConfig) {
        return LocationsApiFp(this.configuration).v11LocationsEntityIdThingsGet(entityId, $skip, $top, $count, $select, $expand, $filter, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Returns all the locations provided by this api (subject to any parameters set)
     * @summary Get all locations
     * @param {number} [$skip] The number of elements to skip from the collection
     * @param {number} [$top] The number of elements to return
     * @param {boolean} [$count] Flag indicating if the total number of items in the collection should be returned.
     * @param {string} [$select] The list of properties that need to be returned
     * @param {string} [$expand] The list of related queries that need to be included in the result
     * @param {string} [$filter] A filter query
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LocationsApi
     */
    public v11LocationsGet($skip?: number, $top?: number, $count?: boolean, $select?: string, $expand?: string, $filter?: string, options?: AxiosRequestConfig) {
        return LocationsApiFp(this.configuration).v11LocationsGet($skip, $top, $count, $select, $expand, $filter, options).then((request) => request(this.axios, this.basePath));
    }
}